Pushed dual-implementation case filtering tests down.
Pushed previously-generic test cases for fetching filtered cases down
from abstract test to the test for the (now only) implementation class.
